Refactored icinga messages to a seperate library
